moved get_api to utils.py
