+++ src/corva/application.py
@@ -1,39 +1,13 @@
 from typing import Any, Callable, List, Optional
 
-from corva.api import Api
-from corva.configuration import SETTINGS, Settings
+from corva import utils
+from corva.configuration import SETTINGS
 from corva.models.scheduled import RawScheduledEvent, ScheduledContext
 from corva.models.stream.raw import RawStreamEvent
 from corva.models.task import RawTaskEvent, TaskContext
 from corva.runners.scheduled import scheduled_runner
 from corva.runners.stream import stream_runner
 from corva.runners.task import task_runner
-
-
-def get_api(context: Any, settings: Settings, timeout: Optional[int] = None) -> Api:
-    """Returns Api instance.
-
-    Args:
-        context: AWS Lambda context.
-        settings:
-        timeout: custom api timeout.
-
-    Raises:
-          Exception: if cound not find an api key in the context.
-    """
-
-    try:
-        api_key = context.client_context.env["API_KEY"]
-    except (AttributeError, KeyError):
-        raise Exception('No API Key found.')
-
-    return Api(
-        api_url=settings.API_ROOT_URL,
-        data_api_url=settings.DATA_API_ROOT_URL,
-        api_key=api_key,
-        app_name=settings.APP_NAME,
-        timeout=timeout,
-    )
 
 
 class Corva:
@@ -59,7 +33,7 @@ class Corva:
             cache_settings: additional cache settings.
         """
 
-        self.api = get_api(context=context, settings=SETTINGS, timeout=timeout)
+        self.api = utils.get_api(context=context, settings=SETTINGS, timeout=timeout)
         self.cache_settings = cache_settings or {}
 
     def stream(
