+++ src/corva/utils.py
@@ -1,5 +1,6 @@
-from typing import Optional
+from typing import Any, Optional
 
+from corva import Api
 from corva.configuration import Settings
 from corva.state.redis_adapter import RedisAdapter
 from corva.state.redis_state import RedisState
@@ -40,3 +41,29 @@ def get_cache(
     )
 
     return RedisState(redis=redis_adapter)
+
+
+def get_api(context: Any, settings: Settings, timeout: Optional[int] = None) -> Api:
+    """Returns Api instance.
+
+    Args:
+        context: AWS Lambda context.
+        settings:
+        timeout: custom api timeout.
+
+    Raises:
+          Exception: if cound not find an api key in the context.
+    """
+
+    try:
+        api_key = context.client_context.env["API_KEY"]
+    except (AttributeError, KeyError):
+        raise Exception('No API Key found.')
+
+    return Api(
+        api_url=settings.API_ROOT_URL,
+        data_api_url=settings.DATA_API_ROOT_URL,
+        api_key=api_key,
+        app_name=settings.APP_NAME,
+        timeout=timeout,
+    )
